.vercel - Copia/cache/index_py/scripts/check_neon_tables.py
import psycopg2

PG_URL = "postgresql://neondb_owner:npg_TBWgl4SM1Ejn@ep-morning-water-acsrbm4u-pooler.sa-east-1.aws.neon.tech/neondb?sslmode=require"

def list_tables():
    try:
        conn = psycopg2.connect(PG_URL)
        cursor = conn.cursor()
        print("Checking tables in Neon PostgreSQL...")
        cursor.execute("SELECT table_name FROM information_schema.tables WHERE table_schema = 'public'")
        tables = [row[0] for row in cursor.fetchall()]
        print(f"Tables found: {tables}")
        conn.close()
    except Exception as e:
        print(f"Error: {e}")

if __name__ == "__main__":
    list_tables()
